Rename PipelineMetadataStore
